requirements.txt
asgiref
Django
django-crispy-forms
Pillow
pytz
sqlparse
djangorestframework
